Factor registry into metrics/spans/shared
