+++ exporters/autoexport/registry.go
@@ -18,20 +18,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
 	"sync"
-
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc"
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
-	"go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/trace"
 )
 
-const (
-	otelExporterOTLPProtoEnvKey = "OTEL_EXPORTER_OTLP_PROTOCOL"
-)
+const otelExporterOTLPProtoEnvKey = "OTEL_EXPORTER_OTLP_PROTOCOL"
 
 // registry maintains a map of exporter names to exporter factories
 // func(context.Context) (T, error) that is safe for concurrent use by multiple
@@ -41,35 +31,7 @@ type registry[T any] struct {
 	names map[string]func(context.Context) (T, error)
 }
 
-func newSpanExporterRegistry() registry[trace.SpanExporter] {
-	return registry[trace.SpanExporter]{
-		names: map[string]func(context.Context) (trace.SpanExporter, error){
-			"":     buildOTLPSpanExporter,
-			"otlp": buildOTLPSpanExporter,
-			"none": func(ctx context.Context) (trace.SpanExporter, error) { return noop{}, nil },
-		},
-	}
-}
-
-func newMetricReaderRegistry() registry[metric.Reader] {
-	return registry[metric.Reader]{
-		names: map[string]func(context.Context) (metric.Reader, error){
-			"":     buildOTLPMetricReader,
-			"otlp": buildOTLPMetricReader,
-			"none": func(ctx context.Context) (metric.Reader, error) { return metric.NewManualReader(), nil },
-		},
-	}
-}
-
 var (
-	// spanExporterRegistry is the package level registry of exporter registrations
-	// and their mapping to a SpanExporter factory func(context.Context) (trace.SpanExporter, error).
-	spanExporterRegistry = newSpanExporterRegistry()
-
-	// metricReaderRegistry is the package level registry of exporter registrations
-	// and their mapping to a MetricReader factory func(context.Context) (metric.MetricReader, error).
-	metricReaderRegistry = newMetricReaderRegistry()
-
 	// errUnknownExporter is returned when an unknown exporter name is used in
 	// the OTEL_*_EXPORTER environment variables.
 	errUnknownExporter = errors.New("unknown exporter")
@@ -114,112 +76,4 @@ func (r *registry[T]) drop(key string) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 	delete(r.names, key)
-}
-
-// RegisterSpanExporter sets the SpanExporter factory to be used when the
-// OTEL_TRACES_EXPORTERS environment variable contains the exporter name. This
-// will panic if name has already been registered.
-func RegisterSpanExporter(name string, factory func(context.Context) (trace.SpanExporter, error)) {
-	if err := spanExporterRegistry.store(name, factory); err != nil {
-		// registry.store will return errDuplicateRegistration if name is already
-		// registered. Panic here so the user is made aware of the duplicate
-		// registration, which could be done by malicious code trying to
-		// intercept cross-cutting concerns.
-		//
-		// Panic for all other errors as well. At this point there should not
-		// be any other errors returned from the store operation. If there
-		// are, alert the developer that adding them as soon as possible that
-		// they need to be handled here.
-		panic(err)
-	}
-}
-
-// RegisterMetricReader sets the MetricReader factory to be used when the
-// OTEL_METRICS_EXPORTERS environment variable contains the exporter name. This
-// will panic if name has already been registered.
-func RegisterMetricReader(name string, factory func(context.Context) (metric.Reader, error)) {
-	if err := metricReaderRegistry.store(name, factory); err != nil {
-		// registry.store will return errDuplicateRegistration if name is already
-		// registered. Panic here so the user is made aware of the duplicate
-		// registration, which could be done by malicious code trying to
-		// intercept cross-cutting concerns.
-		//
-		// Panic for all other errors as well. At this point there should not
-		// be any other errors returned from the store operation. If there
-		// are, alert the developer that adding them as soon as possible that
-		// they need to be handled here.
-		panic(err)
-	}
-}
-
-// spanExporter returns a span exporter using the passed in name
-// from the list of registered SpanExporters. Each name must match an
-// already registered SpanExporter. A default OTLP exporter is registered
-// under both an empty string "" and "otlp".
-// An error is returned for any unknown exporters.
-func spanExporter(ctx context.Context, name string) (trace.SpanExporter, error) {
-	exp, err := spanExporterRegistry.load(ctx, name)
-	if err != nil {
-		return nil, err
-	}
-	return exp, nil
-}
-
-// metricReader returns a metric reader using the passed in name
-// from the list of registered metric.Readers. Each name must match an
-// already registered metric.Reader. A default OTLP exporter is registered
-// under both an empty string "" and "otlp".
-// An error is returned for any unknown exporters.
-func metricReader(ctx context.Context, name string) (metric.Reader, error) {
-	exp, err := metricReaderRegistry.load(ctx, name)
-	if err != nil {
-		return nil, err
-	}
-	return exp, nil
-}
-
-// buildOTLPSpanExporter creates an OTLP span exporter using the environment variable
-// OTEL_EXPORTER_OTLP_PROTOCOL to determine the exporter protocol.
-// Defaults to http/protobuf protocol.
-func buildOTLPSpanExporter(ctx context.Context) (trace.SpanExporter, error) {
-	proto := os.Getenv(otelExporterOTLPProtoEnvKey)
-	if proto == "" {
-		proto = "http/protobuf"
-	}
-
-	switch proto {
-	case "grpc":
-		return otlptracegrpc.New(ctx)
-	case "http/protobuf":
-		return otlptracehttp.New(ctx)
-	default:
-		return nil, errInvalidOTLPProtocol
-	}
-}
-
-// buildOTLPMetricReader creates an OTLP metric reader using the environment variable
-// OTEL_EXPORTER_OTLP_PROTOCOL to determine the exporter protocol.
-// Defaults to http/protobuf protocol.
-func buildOTLPMetricReader(ctx context.Context) (metric.Reader, error) {
-	proto := os.Getenv(otelExporterOTLPProtoEnvKey)
-	if proto == "" {
-		proto = "http/protobuf"
-	}
-
-	switch proto {
-	case "grpc":
-		r, err := otlpmetricgrpc.New(ctx)
-		if err != nil {
-			return nil, err
-		}
-		return metric.NewPeriodicReader(r), nil
-	case "http/protobuf":
-		r, err := otlpmetrichttp.New(ctx)
-		if err != nil {
-			return nil, err
-		}
-		return metric.NewPeriodicReader(r), nil
-	default:
-		return nil, errInvalidOTLPProtocol
-	}
 }
